fix(core): address round 20 review blockers
- Add REDIS_PASSWORD_FILE support for Docker secrets pattern
- Return errors from LoadManagerFromEnv/LoadMonitorFromEnv instead of silent fallback
- Add json.Valid check for task config before processing
- Refresh slot config TTL on restart to prevent expiration
- Persist monitor failure counters in Redis with TTL instead of in-memory maps
- Add explicit Redis client timeouts (dial/read/write)
- Switch Containerfile from scratch to distroless/static:nonroot
- Change extractCheckInterval to return int64
- Add port boundary and password file tests

+++ internal/worker/slot.go
@@ -190,7 +190,7 @@ func (s *Slot) run(parent context.Context) {
 
 			return
 		case <-ticker.C:
-			s.tick(ctx, &lastCheck, int64(checkInterval))
+			s.tick(ctx, &lastCheck, checkInterval)
 		}
 	}
 }
@@ -260,11 +260,11 @@ func (s *Slot) backoff(ctx context.Context) {
 	}
 }
 
-func extractCheckInterval(config json.RawMessage, log *slog.Logger) int {
+func extractCheckInterval(config json.RawMessage, log *slog.Logger) int64 {
 	var cfg struct {
 		Bot struct {
-			CheckInterval int `json:"checkInterval"`
-			Interval      int `json:"interval"`
+			CheckInterval int64 `json:"checkInterval"`
+			Interval      int64 `json:"interval"`
 		} `json:"bot"`
 	}
